Rename `see_backup_model` to `seed_backup`
