+++ lib/wallet/wallet_dashboard.dart
@@ -4,7 +4,7 @@ import 'package:go_router/go_router.dart';
 import 'package:ten_ten_one/balance.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading.dart';
 import 'package:ten_ten_one/model/balance_change_notifier.dart';
-import 'package:ten_ten_one/model/seed_backup_model.dart';
+import 'package:ten_ten_one/model/seed_backup_change_notifier.dart';
 import 'package:ten_ten_one/model/service.dart';
 import 'package:ten_ten_one/payment_history_change_notifier.dart';
 import 'package:ten_ten_one/wallet/deposit.dart';
